backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
pandas
python-dotenv
xgboost
scikit-learn
pydantic